+++ tests/e2e/port_config.js
@@ -29,4 +29,23 @@ describe("port directive configuration", function () {
 			});
 		});
 	});
+
+	describe("Set port 8100 on enviroment variable MM_PORT", function () {
+		before(function() {
+			process.env.MM_PORT = 8100;
+			// Set config sample for use in this test
+			process.env.MM_CONFIG_FILE = "tests/configs/port_8090.js";
+		});
+
+		after(function(){
+			delete process.env.MM_PORT;
+		});
+		it("should return 200", function (done) {
+			request.get("http://localhost:8100", function (err, res, body) {
+				expect(res.statusCode).to.equal(200);
+				done();
+			});
+		});
+	});
+
 });